Finishing Product command's & Queries

+++ ElectronicsEshop.Application/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -7,14 +7,20 @@ using Microsoft.Extensions.Logging;
 namespace ElectronicsEshop.Application.Products.Commands.DeleteProduct;
 
 public sealed class DeleteProductCommandHandler(ILogger<DeleteProductCommandHandler> logger,
-    IProductRepository productRepository) : IRequestHandler<DeleteProductCommand>
+    IProductRepository productRepository,
+    IOrderItemRepository orderItemRepository,
+    ICartItemRepository cartItemRepository) : IRequestHandler<DeleteProductCommand>
 {
     public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
         var product = await productRepository.GetByIdAsync(request.Id, cancellationToken)
             ?? throw new NotFoundException(nameof(Product), request.Id);
 
-        //ověřit v OrderRepository / OrderItemRepository, že neexistuje záznam, který je vázaný prodle ProductId v tabulce OrderItem, protože tam mám Restrict mazání !
+        if(await orderItemRepository.ExistsForProductAsync(product.Id, cancellationToken))
+            throw new DomainException($"Produkt nelze smazat – je použit v objednávkách.");
+
+        if(await cartItemRepository.ExistsForProductAsync(product.Id, cancellationToken))
+            throw new DomainException($"Produkt nelze smazat – mají ho uživatelé v košíku.");
 
         await productRepository.DeleteAsync(request.Id, cancellationToken);
         logger.LogInformation("Produkt {ProductId} ({ProductName}) byl odstraněn.", product.Id, product.Name);
